Continue with refactor of existing sector classes (Element, Shape, Route).
Route class no longer has name parameter. Re-implemented SectorElement deserialize method to read all data from the geoJSON. This included removing fields from geoJSON that are not necessary to this and adding methods to the parser to enable this. Also added a new PolygonShape clas for this. Updated sector and parser tests (but not all passing and more to be added for new methods).